+++ lib/extensions/invitation/phoenix/mailers/mailer_view.ex
@@ -1,0 +1,4 @@
+defmodule PowInvitation.Phoenix.MailerView do
+  @moduledoc false
+  use Pow.Phoenix.Mailer.View
+end