+++ lib/extensions/invitation/phoenix/views/invitation_view.ex
@@ -1,0 +1,4 @@
+defmodule PowInvitation.Phoenix.InvitationView do
+  @moduledoc false
+  use Pow.Phoenix.View
+end